+++ src/Udf/PhysicalPartition.cs
@@ -38,7 +38,7 @@ namespace DiscUtils.Udf
         {
             _partitionNumber = descriptor.PartitionNumber;
             _parentBuffer = buffer;
-            _contentBuffer = new SubBuffer(_parentBuffer, descriptor.PartitionStartingLocation * sectorSize, descriptor.PartitionLength * sectorSize);
+            _contentBuffer = new SubBuffer(_parentBuffer, descriptor.PartitionStartingLocation * (long)sectorSize, descriptor.PartitionLength * (long)sectorSize);
             _descriptor = descriptor;
         }
